pyproject.toml
[tool.poetry]
name = "financebot"
version = "0.1.0"
description = ""
authors = ["stepashka <cociorvastefan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"

[virtualenvs]
in-project = false


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
